ARCO-175: rename callbacker -> sender
